requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
ImageHash
supabase
python-dotenv
python-multipart
Pillow
cryptography
